Persist generated lobby rule profiles
- Generate deterministic rule profiles for configured board and tile-set
sizes
- Scale premium squares and tile quantities from the classic profile
- Embed complete immutable rules in game-start events and recovered
state
- Version journal payloads while decoding legacy games with classic
rules
- Launch lobbies with their selected rules and use persisted rules
during play

+++ packages/app/src/main.rs
@@ -157,8 +157,8 @@ impl RuntimeConfig {
         let max_tile_sets = configured_limit("WWMTF_MAX_TILE_SETS", production_mode, 16)?;
         let game_creation_policy = wwmtf_app::GameCreationPolicy::new(
             max_players,
-            u16::try_from(max_board_size)?,
-            u16::try_from(max_tile_sets)?,
+            u8::try_from(max_board_size)?,
+            u8::try_from(max_tile_sets)?,
         )?;
         let google = google_runtime_config(
             std::env::var("WWMTF_GOOGLE_CLIENT_ID").ok(),
